backend/requirements.txt
Django==2.2.16
djangorestframework==3.12.4
django-filter==21.1
djoser==2.1.0
toml==0.10.2
Pillow==9.5.0
requests==2.26.0
